# api/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "data_channel.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "DataChannelService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/data/file": {
      "post": {
        "operationId": "DataChannelService_UploadFile",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/data_channel_serviceUploadFileResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/data_channel_serviceUploadFileRequest"
            }
          }
        ],
        "tags": [
          "DataChannelService"
        ]
      }
    },
    "/data/{sessionId}/history": {
      "get": {
        "operationId": "DataChannelService_GetHistory",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/data_channel_serviceGetHistoryResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "DataChannelService"
        ]
      }
    }
  },
  "definitions": {
    "data_channel_serviceDataMessage": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "senderId": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      }
    },
    "data_channel_serviceGetHistoryResponse": {
      "type": "object",
      "properties": {
        "messages": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/data_channel_serviceDataMessage"
          }
        }
      }
    },
    "data_channel_serviceUploadFileRequest": {
      "type": "object",
      "properties": {
        "sessionId": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "filename": {
          "type": "string"
        },
        "content": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "data_channel_serviceUploadFileResponse": {
      "type": "object",
      "properties": {
        "fileId": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}